Added input validation to the login page
In order to implement this feature, a refactoring of the UI has been
done to have interior mutability with the state of some components.
Some components had logic that should only be handled by
the component itself internally, which avoids the following issues:
- Duplication of code and states
- Issues when trying to mutate the state of a component
- Also fixed a UI bug that was making switching fields impossible

Input validation checks:
- Checks if the email is valid
- Checks if the password has the following requirements when signing in:
    - 8 characters minimum
    - 1 lowercase letter
    - 1 uppercase letter
    - 1 special symbol
    - No spaces

Furthermore the error messages should be improved to provide a better
user experience.